requirements.txt
numpy
pytorch
matplotlib
scikit-image
jupyter